* renamed headers with implementation to hpp *
