+++ examples/ex27.C
@@ -17,8 +17,8 @@ any standard 3D cross-platform API, so you get this instead.)
 ---------------------------------------------------------------------------- */
 #include <cmath>
 #include <GASStateGA.h>
-#include <GAList.h>
-#include <GA1DArrayGenome.h>
+#include <GAList.hpp>
+#include <GA1DArrayGenome.hpp>
  
 
 #include <iostream>
@@ -27,15 +27,7 @@ any standard 3D cross-platform API, so you get this instead.)
  
  
 
-// force instantiations for compilers that do not do auto instantiation
-// for some compilers (e.g. metrowerks) this must come after any
-// specializations or you will get 'multiply-defined errors when you compile.
-#if !defined(GALIB_USE_AUTO_INST)
-#include <GAList.C>
-#include <GA1DArrayGenome.C>
-GALIB_INSTANTIATION_PREFIX GAList<int>;
-GALIB_INSTANTIATION_PREFIX GA1DArrayGenome<float>;
-#endif
+
